refactor(core): replace star queries/types with pin equivalents
